Fixed casting that did not work in 64 bit.
Because size_t was of the same size as uint32_t in 32 bit mode it worked,
however, in 64 bit builds size_t is greater than uint32_t and as so I'll
need to uppcast numOfRemoves to size_t instead.

+++ Doremi/Core/Source/FrequencyBufferHandler.cpp
@@ -66,7 +66,7 @@ namespace Doremi
             }
 
             // Check if we can remove (error check), then remove the number of accs in difference
-            numOfRemoves = std::min(static_cast<uint32_t>(numOfRemoves), m_bufferedFrequencies.size());
+            numOfRemoves = std::min(static_cast<size_t>(numOfRemoves), m_bufferedFrequencies.size());
 
             // Check if we can remove (error check), then remove the number of accs in difference
             if(numOfRemoves <= m_bufferedFrequencies.size())
